refactor: move AgendaDetailsOneTimeEvent to separate file
